Add network logging with Timber and configure OkHttpClient; update API key handling and permissions

+++ domain/src/main/java/com/velosobr/domain/model/Exchange.kt
@@ -16,5 +16,5 @@ data class Exchange(
     val volume1hrsUsd: Double,
     val volume1dayUsd: Double,
     val volume1mthUsd: Double,
-    val rank: Int
+    val rank: Double
 )
